Crate mpz_core

Core types and utilities for MPC protocols

Re-exports

pub use block::Block;

pub use block::BlockSerialize;

Modules

aes

Fixed-key AES cipher

block

A block of 128 bits and its operations.

commit

This module provides a hash commitment scheme for types which implement CanonicalSerialize

ggm_tree

Implement GGM tree for OT. Implementation of GGM based on the procedure explained in the write-up (https://eprint.iacr.org/2020/925.pdf, Page 14)

hash

Traits and types for hashing serde serializable types.

lpn

Implement LPN with local linear code. More specifically, a local linear code is a random boolean matrix with at most D non-zero values in each row.

prg

Implement AES-based PRG.

prp

An implementation of Pseudo Random Permutation (PRP) based on AES.

serialize

Traits for canonical serialization of serde serializable types.

tkprp

Implement the two-key PRG as G(k) = PRF_seed0(k)\xor k || PRF_seed1(k)\xor k Refer to (https://www.usenix.org/system/files/conference/nsdi17/nsdi17-wang-frank.pdf, Page 8)

utils

Utilities for MPC protocols

Macros

impl_domain_separated_hash

A macro for implementing the DomainSeparatedHash trait

Traits

ProtocolMessage

A protocol with a message type.