mpz_core::block
Struct Block

pub struct Block(/* private fields */);

A block of 128 bits

Implementations

impl Block

pub const LEN: usize = 16usize

The length of a block in bytes

pub const ZERO: Self = _

A zero block

pub const ONE: Self = _

A one block

pub const ONES: Self = _

A block with all bits set to 1

pub const SELECT_MASK: [Self; 2] = _

A length 2 array of zero and one blocks

pub fn new(bytes: [u8; 16]) -> Self

Create a new block

pub fn to_bytes(self) -> [u8; 16]

Returns the byte representation of the block

pub fn random<R: Rng + CryptoRng + ?Sized>(rng: &mut R) -> Self

Generate a random block using the provided RNG

pub fn random_array<const N: usize, R: Rng + CryptoRng>( rng: &mut R, ) -> [Self; N]

Generate a random array of blocks using the provided RNG

pub fn random_vec<R: Rng + CryptoRng + ?Sized>( rng: &mut R, n: usize, ) -> Vec<Self>

Generate a random vector of blocks using the provided RNG

pub fn clmul(self, other: Self) -> (Self, Self)

Carry-less multiplication of two blocks, without the reduction step.

pub fn reduce_gcm(x: Self, y: Self) -> Self

Reduces the polynomial represented in bits modulo the GCM polynomial x^128 + x^7 + x^2 + x + 1. x and y are resp. upper and lower bits of the polynomial.

pub fn gfmul(self, x: Self) -> Self

The multiplication of two Galois field elements.

pub fn inn_prdt_no_red(a: &[Block], b: &[Block]) -> (Block, Block)

Compute the inner product of two block vectors, without reducing the polynomial.

pub fn inn_prdt_red(a: &[Block], b: &[Block]) -> Block

Compute the inner product of two block vectors.

pub fn reverse_bits(self) -> Self

Reverses the bits of the block

pub fn set_lsb(&mut self)

Sets the least significant bit of the block

pub fn lsb(&self) -> usize

Returns the least significant bit of the block

pub fn sigma(a: Self) -> Self

Let x0 and x1 be the lower and higher halves of x, respectively. This function compute sigma( x = x0 || x1 ) = x1 || (x0 xor x1).